Exception

+++ src/main/java/pl/ppyrczak/busschedulesystem/service/ReviewService.java
@@ -2,6 +2,9 @@ package pl.ppyrczak.busschedulesystem.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import pl.ppyrczak.busschedulesystem.exception.ApiRequestException;
+import pl.ppyrczak.busschedulesystem.exception.IllegalDateException;
+import pl.ppyrczak.busschedulesystem.exception.IllegalPassengerException;
 import pl.ppyrczak.busschedulesystem.model.Passenger;
 import pl.ppyrczak.busschedulesystem.model.Review;
 import pl.ppyrczak.busschedulesystem.model.Schedule;
@@ -26,20 +29,25 @@ public class ReviewService {
 
     public Review addReview(Review review) {
         review.setCreated();
-        if (!checkIfReviewIsNotBeforeArrival(review)) {
-            throw new RuntimeException("You can not add review before arrival");
-        } else if (!checkIfScheduleHasPassenger(review)) {
-            throw new RuntimeException("passenger does not exist");
+        Schedule schedule = scheduleRepository.findById(review.getScheduleId())
+                .orElseThrow(() -> new ApiRequestException(
+                        "Schedule with id " + review.getScheduleId() + " does not exist"
+                ));
+        if (!reviewIsNotBeforeArrival(review)) {
+            throw new IllegalDateException(schedule.getArrival());
+        } else if (!scheduleHasPassenger(review)) {
+            throw new IllegalPassengerException();
         } else {
             return reviewRepository.save(review);
         }
-
     }
 
     public void deleteReview(Long id) {
+        if (!reviewRepository.findById(id).isPresent()) {
+            throw new ApiRequestException("Review with id " + id + " does not exist");
+        }
         reviewRepository.deleteById(id);
     }
-
 
     public List<Review> getReviewsWithDetailsForSpecificSchedule(Long id) {
         List<Review> reviews = reviewRepository.
@@ -47,20 +55,18 @@ public class ReviewService {
         return reviews;
     }
 
-    private boolean checkIfReviewIsNotBeforeArrival(Review review) { //chyba dziala
+    private boolean reviewIsNotBeforeArrival(Review review) {
         boolean returnStat = true;
         Long scheduleId = review.getScheduleId();
         Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow();
         if (review.getCreated().isBefore(schedule.getArrival())) {
             returnStat = false;
         }
-
         return returnStat;
     }
 
-    private boolean checkIfScheduleHasPassenger(Review review) {
+    private boolean scheduleHasPassenger(Review review) {
         boolean returnStat = false;
-
         Long scheduleId = review.getScheduleId();
         Long passengerId = review.getPassengerId();
         Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow();
